added turbines controller, get/add turbine methods, get optimal turbine method, calc turbine rgc method

+++ server/models/models.js
@@ -24,7 +24,7 @@ const Turbine = sequelize.define('turbine', {
     mark: {type: DataTypes.STRING},
     ['electricity_power']: {type: DataTypes.INTEGER},
     ['thermal_power']: {type: DataTypes.INTEGER},
-    ['power_generation']: {type: DataTypes.INTEGER}
+    ['power_generation']: {type: DataTypes.FLOAT}
 })
 
 // Годовой план станции
